Move music browser package
