+++ app/src/main/java/com/marverenic/music/ui/browse/FolderSection.java
@@ -1,29 +1,28 @@
-package com.marverenic.music.ui.library.browse;
+package com.marverenic.music.ui.browse;
 
 import android.databinding.DataBindingUtil;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
 import com.marverenic.adapter.EnhancedViewHolder;
 import com.marverenic.adapter.HeterogeneousAdapter;
 import com.marverenic.music.R;
-import com.marverenic.music.databinding.InstanceFileBinding;
+import com.marverenic.music.databinding.InstanceFolderBinding;
 
 import java.io.File;
 import java.util.List;
 
-public class FileSection extends HeterogeneousAdapter.ListSection<File> {
+public class FolderSection extends HeterogeneousAdapter.ListSection<File> {
 
-    private ThumbnailLoader mThumbnailLoader;
-    private FileViewModel.OnFileSelectedListener mSelectionListener;
+    @Nullable
+    private FolderViewModel.OnFolderSelectedListener mSelectionListener;
 
-    public FileSection(@NonNull List<File> data,
-                       @NonNull FileViewModel.OnFileSelectedListener selectionListener,
-                       @NonNull ThumbnailLoader thumbnailLoader) {
+    public FolderSection(@NonNull List<File> data,
+                         @Nullable FolderViewModel.OnFolderSelectedListener selectionListener) {
         super(data);
         mSelectionListener = selectionListener;
-        mThumbnailLoader = thumbnailLoader;
     }
 
     @Override
@@ -35,28 +34,26 @@ public class FileSection extends HeterogeneousAdapter.ListSection<File> {
     public EnhancedViewHolder<File> createViewHolder(HeterogeneousAdapter adapter, ViewGroup parent) {
         LayoutInflater inflater = LayoutInflater.from(parent.getContext());
         return new ViewHolder(
-                DataBindingUtil.inflate(inflater, R.layout.instance_file, parent, false));
+                DataBindingUtil.inflate(inflater, R.layout.instance_folder, parent, false));
     }
 
     private class ViewHolder extends EnhancedViewHolder<File> {
 
-        private InstanceFileBinding mBinding;
-        private FileViewModel mViewModel;
+        private InstanceFolderBinding mBinding;
+        private FolderViewModel mViewModel;
 
-        public ViewHolder(InstanceFileBinding binding) {
+        public ViewHolder(InstanceFolderBinding binding) {
             super(binding.getRoot());
             mBinding = binding;
-            mViewModel = new FileViewModel(itemView.getContext(), mThumbnailLoader);
-            mViewModel.setFileSelectionListener(mSelectionListener);
-
+            mViewModel = new FolderViewModel();
+            mViewModel.setSelectionListener(mSelectionListener);
             mBinding.setViewModel(mViewModel);
         }
 
         @Override
         public void onUpdate(File item, int position) {
-            mViewModel.setFile(item);
+            mViewModel.setFolder(item);
             mBinding.executePendingBindings();
         }
     }
-
 }
